Implement AddPid method for cgroup managers
To this day, cgroup managers did not have an ability to add a process
to an existing cgroup. One might use cgroups.WriteCgroupProc, but
this is cgroupfs operation and does not take into account systemd.

Let's introduce AddPid, which will add a process, identified by PID,
to the cgroup (or, optionally, a sub cgroup) of the manager.

This will allow runtimes like runc to ask systemd to move the process
into a proper unit, instead of using cgroupfs directly.

Implementation for systemd requires [1] and systemd >= v238.

[1]: https://github.com/coreos/go-systemd/pull/458
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>